Refactor database config
